+++ resources/lang/en/route-descriptions.php
@@ -19,7 +19,8 @@ return [
     'auth_permitted_resource_types_GET' => 'Return the permitted resource types for the signed-in user',
     'auth_resource_GET' => 'Return the selected resource for the signed-in user',
     'auth_resources_GET' => 'Return the resources for the selected permitted resource type for the signed-in user',
-    'auth_request_resource_delete_POST' => 'Request the deletion of a resource all all dependant data, this is and unrecoverable action',
+    'auth_request_resource_delete_POST' => 'Request the deletion of a resource and all dependant data, this is and unrecoverable action',
+    'auth_request_resource_type_delete_POST' => 'Request the deletion of a resource type and all dependant data, this is and unrecoverable action',
 
     'api_GET_index' => 'Return all the API routes',
     'api_GET_changelog' => 'Return the complete API changelog',